feat: use two separate pages and routes for Terms of Service and Privacy Policy
