+++ packages/pn-pa-webapp/src/navigation/routes.tsx
@@ -5,10 +5,11 @@ import Dashboard from '../pages/Dashboard.page';
 import NotificationDetail from '../pages/NotificationDetail.page';
 import NewNotification from '../pages/NewNotification.page';
 // import ApiKeys from '../pages/ApiKeys.page';
-import PrivacyTOSPage from "../pages/PrivacyTOS.page";
+import PrivacyPolicyPage from "../pages/PrivacyPolicy.page";
+import TermsOfServicePage from "../pages/TermsOfService.page";
 
 import { PNRole } from '../models/user';
-import TermsOfService from '../pages/TermsOfService.page';
+import ToSAcceptance from '../pages/ToSAcceptance.page';
 import * as routes from './routes.const';
 import SessionGuard from './SessionGuard';
 import RouteGuard from './RouteGuard';
@@ -19,7 +20,7 @@ function Router() {
       <Route path="/" element={<SessionGuard />}>
         {/* protected routes */}
         <Route path="/"  element={<RouteGuard roles={[PNRole.ADMIN, PNRole.OPERATOR]} />}>
-          <Route path={routes.TOS} element={<TermsOfService />} />
+          <Route path={routes.TOS} element={<ToSAcceptance />} />
           <Route path={routes.DASHBOARD} element={<Dashboard />} />
           <Route path={routes.DETTAGLIO_NOTIFICA} element={<NotificationDetail />} />
           <Route path={routes.NUOVA_NOTIFICA} element={<NewNotification />} />
@@ -41,8 +42,8 @@ function Router() {
       </Route>
       <Route path="*" element={<NotFound />} />
       {/* PN-2391 element will be fixed in PN-2394 */}
-      <Route path={routes.PRIVACY_POLICY} element={<PrivacyTOSPage />} />
-      <Route path={routes.TERMS_OF_SERVICE} element={<PrivacyTOSPage />} />
+      <Route path={routes.PRIVACY_POLICY} element={<PrivacyPolicyPage />} />
+      <Route path={routes.TERMS_OF_SERVICE} element={<TermsOfServicePage />} />
     </Routes>
   );
 }
