+++ packages/pn-pa-webapp/src/pages/TermsOfService.page.tsx
@@ -1,53 +1,168 @@
-import { Fragment, ReactNode, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { Trans } from 'react-i18next';
-import { Grid } from '@mui/material';
-import { TermsOfServiceHandler, useIsMobile } from '@pagopa-pn/pn-commons';
+import {
+  Box, Container,
+  Link,
+  Typography
+} from "@mui/material";
 
-import { useAppDispatch, useAppSelector } from '../redux/hooks';
-import { acceptToS } from '../redux/auth/actions';
-import * as routes from '../navigation/routes.const';
-import { RootState } from '../redux/store';
+const PrivacyPolicyPage = () => (
+    <>
+      <Container
+        maxWidth="xl"
+        sx={{py: 4}}
+      >
+        <Typography variant="h2">Termini e condizioni d’uso</Typography>
+        <Typography variant="h4">Piattaforma Notifiche Digitali</Typography>
+        <Typography variant="h5">Portale back office “Area Riservata”</Typography>
+        <Box mt={3}>
+          <Typography variant="body1">
+            <strong>
+              I presenti termini e condizioni d’uso (di seguito anche “ToS”) regolano il rapporto tra gli utenti del
+              sito <Link href="https://pn.selfcare.pagopa.it">pn.selfcare.pagopa.it</Link> (di seguito “Sito”) dove
+              poggia parte della Piattaforma Notifiche Digitali (di seguito “Piattaforma”) prevista dall’art. 26,
+              comma 2, lettera c), del decreto-legge n. 76 del 17 luglio 2020 come convertito, con modificazioni, dalla
+              legge 11 settembre 2020, n. 120 e come altresì modificato dal decreto-legge. n. 77 del 31 maggio 2021,
+              così come convertito dalla legge 29 luglio 2021 n. 108 e gestita da PagoPA S.p.A. (di seguito “Società”)
+              stabilendo diritti ed obblighi nell’utilizzo dello stesso.<br/><br/>
+              Leggili attentamente: utilizzando il Sito, li accetti integralmente e ti impegni a rispettarli.
+            </strong>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            1. Descrizione del servizio
+          </Typography>
+          <Typography variant="body1">
+            1.1 PagoPA S.p.A. con sede legale in Roma, Piazza Colonna 370, CAP 00187, n. di iscrizione a Registro
+            Imprese di Roma, CF e P.IVA 15376371009, in base a quanto previsto dall’art. 8, commi 2 e 3, del D.L. n.
+            135/2018, ha sviluppato il Sito a cui si accede a mezzo del portale di back office <b>“Area Riservata”</b> (di
+            seguito anche il <b>“Portale”</b>) della Società, raggiungibile all’indirizzo web
+            <Link href="https://selfcare.pagopa.it">https://selfcare.pagopa.it</Link>.<br/>
+            1.2 Mediante il Sito, le pubbliche amministrazioni, ovvero le amministrazioni individuate dall’articolo
+            26, comma 2, lettera c), del decreto-legge n. 76 del 17 luglio 2020 come convertito, con modificazioni,
+            dalla legge 11 settembre 2020, n. 120 e come altresì modificato dal decreto-legge. n. 77 del 31 maggio
+            2021, così come convertito dalla legge 29 luglio 2021 n. 108 (di seguito anche <b>“PA Mittenti”</b>) aderiscono
+            alla Piattaforma per il tramite dei propri referenti (di seguito <b>“Utenti”</b>).
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            2. Costi
+          </Typography>
+          <Typography variant="body1">
+            2.1 L’utilizzo del Sito non prevede il pagamento di alcun corrispettivo da parte dell’Utente.<br/>
+            2.2 Per utilizzare il Sito è necessaria una connessione dati i cui costi non sono a carico della Società.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            3. Siti di terze parti
+          </Typography>
+          <Typography variant="body1">
+            3.1 Tutti i siti web che non appartengono alla Società e a cui l’Utente può accedere, anche a seguito di accesso a collegamento ipertestuale (link) presente sul Sito, sono indipendenti dal Sito stesso. La Società non ha alcun potere di controllo, né diretto né indiretto, sugli stessi e sul loro contenuto o utilizzo.
+            <br/>
+            3.2 L’Utente riconosce e accetta che nessun tipo di responsabilità, neanche indiretta, potrà essere imputata alla Società per contenuti, prodotti o altri beni/servizi, nonché per l’utilizzazione dei collegamenti ipertestuali di cui sopra, e che la presenza sul Sito di un collegamento ipertestuale (link) verso un altro indirizzo web non comporta alcun tipo di assunzione o accettazione di responsabilità da parte della Società medesima circa il contenuto o l’utilizzazione del sito collegato. L’Utente accetta espressamente, quindi, di manlevare e tenere indenne la Società da ogni danno, perdita e conseguenza pregiudizievole patita dall’Utente.
+            <br/>
+            3.3 La disciplina in materia di protezione dei dati personali stabilita nel presente Sito non trova applicazione con riguardo a siti di terze parti su cui la Società non esercita alcun controllo. La Società non potrà essere, quindi, ritenuta responsabile per l’eventuale violazione della normativa in materia di protezione dei dati personali posta in essere da tali siti di terze parti.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            4. Assistenza all’Utente e debug
+          </Typography>
+          <Typography variant="body1">
+            4.1 La Società fa del suo meglio per assicurare un’esperienza adatta alle esigenze dell’Utente.
+            <br/>
+            4.2 Se l’Utente ha bisogno di assistenza sul Sito, è possibile accedere alla sezione “?Assistenza” cliccando sull’apposita sezione in alto a destra della pagina ed aprire un ticket. Per maggiori informazioni su come trattiamo i dati personali in caso di assistenza, si prega di prendere visione dell’Informativa Privacy Assistenza presente sul sito <Link href="https://www.pagopa.it">www.pagopa.it</Link> al seguente link: <Link href="https://www.pagopa.it/it/privacy-policy-assistenza/">https://www.pagopa.it/it/privacy-policy-assistenza/</Link>.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            5. Responsabilità e obblighi dell’Utente
+          </Typography>
+          <Typography variant="body1" component="span">
+            <ul style={{listStyleType: "lower-alpha"}}>
+              <li>
+                <i>Uso conforme e obblighi dell’Utente</i>
+                <br/>
+                Utilizzando il Sito, l’Utente accetta i presenti ToS e si impegna a rispettarli in ogni momento.
+                L’Utente può utilizzare il Sito esclusivamente per gli scopi autorizzati dai presenti ToS e in
+                conformità alla legge applicabile.<br/>
+                L’Utente si impegna a non utilizzare, direttamente o indirettamente, il Sito:
+                <ul style={{listStyleType: "lower-roman"}}>
+                  <li>in violazione dei diritti della Società o di terzi;</li>
+                  <li>per finalità di marketing, promozionali e/o pubblicitarie;</li>
+                  <li>in modo illecito, diffamatorio, osceno, volgare, intimidatorio, offensivo nei confronti della Società o di terzi;</li>
+                  <li>per trasmettere comunicazioni illecite o non autorizzate, quali messaggi massivi, spam o messaggi automatici;</li>
+                  <li>per comunicare dati o informazioni false, ingannevoli o illecite;</li>
+                  <li>per trasmettere virus, malware o altri codici dannosi per qualsivoglia dispositivo o sistema e/o effettuare interventi di hacking;</li>
+                  <li>aggirare o manomettere l’accesso e l’autenticazione;</li>
+                  <li>accedere attraverso programmi o metodi diversi da quelli ufficialmente rilasciati e gestiti dalla Società;</li>
+                  <li>farne un uso in violazione o sovraccarico della capacità di rete.</li>
+                </ul>
+              </li>
+              <li>
+                <i>Responsabilità Utente, indennizzo e manleva</i>
+                <br/>
+                L’Utente è tenuto ad avvisare tempestivamente la Società nel caso venga a conoscenza di uso o accesso non autorizzato al Sito o di qualsiasi violazione, malfunzionamento o incidente di sicurezza. Qualora tali eventi comportino violazioni e intrusioni nei dati personali, l’Utente ne informerà tempestivamente la Società entro 24 ore dal verificarsi dell’evento. La Società non potrà essere considerata responsabile per eventuali danni o disservizi derivanti da usi ed accessi non autorizzati effettuati tramite l’utilizzo delle credenziali in possesso dell’Utente.
+                <br/>
+                Fermi restando gli altri rimedi a disposizione della Società ai sensi della normativa applicabile, nel caso di violazioni ripetute, gravi o sostanziali dei ToS la Società si riserva il diritto di sospendere o disattivare permanentemente l’utenza del responsabile della violazione.
+                <br/>
+                Inoltre, l’Utente riconosce che le credenziali di autenticazione sono strettamente personali, devono essere mantenute riservate e non sono condivisibili con terzi, se non negli eventuali casi espressamente previsti dalla legge.
+                <br/>
+                L’Utente si obbliga a manlevare e tenere indenne la Società e i suoi dipendenti, dirigenti e amministratori da qualsiasi domanda, richiesta, reclamo, pretesa o azione legale da parte di terzi, nonché da qualsiasi danno, perdita o spesa (incluse le ragionevoli spese legali necessarie per resistere le predette azioni) da chiunque subiti, direttamente o indirettamente, in relazione, in conseguenza o in connessione:
+                <ul style={{listStyleType: "lower-roman"}}>
+                  <li>alle attività e alla condotta dell’Utente nell’utilizzo del Sito e delle sue funzionalità;</li>
+                  <li>alle dichiarazioni, ai dati e alle informazioni comunque fornite dall’Utente;</li>
+                  <li>alle violazioni dei ToS o della legge applicabile da parte dell’Utente.</li>
+                </ul>
+              </li>
+            </ul>
+          </Typography>
+          <Typography variant="body1" fontWeight="bold">
+            6. Misure a disposizione della Società
+          </Typography>
+          <Typography variant="body1" component="span">
+            6.1 Ogni attività effettuata sul Sito da un Utente è eseguita in nome proprio con ogni conseguenza ai fini di legge anche in merito ad eventuali responsabilità per eventuali perdite o danni subiti dalla Società o da terzi in caso di violazioni dei presenti ToS, nonché dell’utilizzo improprio o illecito del Sito. La Società può adottare autonomamente tutte le misure necessarie, inclusa la sospensione o l’interruzione del Sito o dell’utilizzo dello stesso da parte di un Utente, qualora l’Utente ne faccia un utilizzo illecito o altrimenti vietato, oppure commetta una violazione sostanziale o reiterata dei presenti ToS.
+            <br/>
+            6.2 La Società può - in conformità alla normativa primaria e secondaria applicabile - rilevare gli accessi al Sito e laddove riscontrasse una violazione delle procedure di sicurezza e protezione dei dati, sospendere temporaneamente l’utenza e disattivarla, fatto salvo ogni altro rimedio, anche giurisdizionale, consentito dalla legge.
+          </Typography>
+          <Typography variant="body1" fontWeight="bold">
+            7. Esclusioni e limitazione di responsabilità
+          </Typography>
+          <Typography variant="body1">
+            7.1 Gli utenti sono i soli responsabili dei dati inseriti sul Sito. Prima di fornire dati di terzi, si invita l’Utente a verificare che abbia ricevuto ogni necessaria autorizzazione all’uso di tali dati con specifico riferimento alla loro diffusione sul Sito.
+            <br/>
+            7.2 La Società è responsabile unicamente delle funzionalità del Sito e, di conseguenza, non potranno esserle imputati danni o perdite che non siano diretta conseguenza della violazione dei presenti ToS da parte della stessa.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            8. Modifiche ai ToS
+          </Typography>
+          <Typography variant="body1">
+            8.1 La Società cerca di offrire all’Utente nuove funzionalità e di migliorare il Sito. Per questa ragione, i presenti ToS potrebbero cambiare nel tempo e, a tal fine, la Società chiede all’Utente di monitorarne l’eventuale aggiornamento ogni volta in cui accede al Sito. Si ricorda che se l’Utente continuerà ad utilizzare le funzionalità del Sito, accetterà le eventuali modifiche apportate dalla Società ai presenti ToS.
+            <br/>
+            8.2 Se l’Utente non intende accettare le modifiche apportate ai presenti ToS, come unico rimedio, potrà, in qualsiasi momento, per qualsiasi motivo, senza alcun preavviso e senza alcun costo o onere, cessare l’utilizzo del Sito.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            9. Proprietà intellettuale
+          </Typography>
+          <Typography variant="body1">
+            9.1 La Società, nel rispetto dei termini dell’accordo di adesione alla Piattaforma, fornisce all’Utente una licenza limitata e non esclusiva per l’utilizzo del Sito.
+            <br/>
+            9.2 PagoPA S.p.A. è il titolare esclusivo dei contenuti del Sito, ivi compresi, a titolo esemplificativo e non esaustivo, immagini e layout delle pagine, design e il know-how. Alcuni dei predetti contenuti possono essere coperti da copyright, marchi, brevetti, modelli e/o altri diritti di proprietà industriale ed intellettuale riconosciuti dall’ordinamento italiano ed internazionale.
+            <br/>
+            9.3 Nessun contenuto del Sito può essere considerato o interpretato come concesso in licenza e/o come oggetto di qualsivoglia altro diritto di utilizzo da parte degli utenti e/o di soggetti terzi.
+          </Typography>
+          <Typography variant="body1" fontWeight="bold">
+            10. Legge applicabile e foro competente
+          </Typography>
+          <Typography variant="body1">
+            10.1 I presenti ToS sono regolati dalla legge italiana.
+            <br/>
+            10.2 Ogni controversia che dovesse insorgere tra l’Utente e la Società in relazione ai presenti ToS e/o al Sito sarà di competenza esclusiva del foro di Roma oppure del giudice del luogo di residenza o domicilio dell’Utente, ove qualificato come consumatore ai sensi del D.Lgs. 206/2005.
+          </Typography>
+          <br/>
+          <Typography variant="body1">
+            Proseguendo l’Utente dichiara espressamente di aver letto, compreso e accettato i ToS e in particolare i seguenti articoli: <b>(5)</b> Responsabilità e obblighi dell’Utente; <b>(7)</b> Esclusioni e limitazione di responsabilità; <b>(8)</b> Modifiche ai ToS e <b>(10)</b> Legge applicabile e foro competente.
+          </Typography>
+        </Box>
+      </Container>
+    </>);
 
-const TermsOfService = () => {
-  const isMobile = useIsMobile();
-  const dispatch = useAppDispatch();
-  const navigate = useNavigate();
-  const tos = useAppSelector((state: RootState) => state.userState.tos);
-
-  const handleAccept = () => {
-    void dispatch(acceptToS())
-      .unwrap()
-      .then(() => {
-        navigate(routes.DASHBOARD);
-      })
-      .catch((_) => {
-        console.error(_);
-      });
-  };
-
-  useEffect(() => {
-    if (tos) {
-      navigate(routes.DASHBOARD);
-    }
-  }, [tos]);
-
-  return (
-    <Fragment>
-      <Grid container justifyContent="center" my={isMobile ? 4 : 16}>
-        <Grid item xs={10} sm={8} md={4} display="flex" alignItems="center" flexDirection="column">
-          <TermsOfServiceHandler
-            handleAcceptTos={handleAccept}
-            transComponent={(path: string, components: Array<ReactNode>, defaultLoaclization: ReactNode) => (
-              <Trans i18nKey={path} components={components}>
-                {defaultLoaclization}
-              </Trans>
-            )}
-          />
-        </Grid>
-      </Grid>
-    </Fragment>
-  );
-};
-
-export default TermsOfService;
+export default PrivacyPolicyPage;
