+++ packages/pn-personafisica-webapp/src/navigation/routes.tsx
@@ -7,13 +7,14 @@ import SessionGuard from './SessionGuard';
 import RouteGuard from './RouteGuard';
 
 const Profile = React.lazy(() => import('../pages/Profile.page'));
-const TermsOfService = React.lazy(() => import('../pages/TermsOfService.page'));
+const ToSAcceptance = React.lazy(() => import('../pages/ToSAcceptance.page'));
 const Notifiche = React.lazy(() => import('../pages/Notifiche.page'));
 const NotificationDetail = React.lazy(() => import('../pages/NotificationDetail.page'));
 const Contacts = React.lazy(() => import('../pages/Contacts.page'));
 const Deleghe = React.lazy(() => import('../pages/Deleghe.page'));
 const NuovaDelega = React.lazy(() => import('../pages/NuovaDelega.page'));
-const PrivacyTOSPage = React.lazy(() => import('../pages/PrivacyTOS.page'));
+const PrivacyPolicyPage = React.lazy(() => import('../pages/PrivacyPolicy.page'));
+const TermsOfServicePage = React.lazy(() => import('../pages/TermsOfService.page'));
 
 function Router() {
   return (
@@ -22,7 +23,7 @@ function Router() {
         <Route path="/" element={<SessionGuard />}>
           {/* protected routes */}
           <Route path="/" element={<RouteGuard />}>
-            <Route path={routes.TOS} element={<TermsOfService />} />
+            <Route path={routes.TOS} element={<ToSAcceptance />} />
             <Route path={routes.NOTIFICHE} element={<Notifiche />} />
             <Route path={routes.NOTIFICHE_DELEGATO} element={<Notifiche />} />
             <Route path={routes.DETTAGLIO_NOTIFICA} element={<NotificationDetail />} />
@@ -37,9 +38,8 @@ function Router() {
             <Route path="*" element={<NotFound />} />
           </Route>
         </Route>
-        {/* PN-2391 element will be fixed in PN-2394 */}
-        <Route path={routes.PRIVACY_POLICY} element={<PrivacyTOSPage />} />
-        <Route path={routes.TERMS_OF_SERVICE} element={<PrivacyTOSPage />} />
+        <Route path={routes.PRIVACY_POLICY} element={<PrivacyPolicyPage />} />
+        <Route path={routes.TERMS_OF_SERVICE} element={<TermsOfServicePage />} />
         <Route path="*" element={<NotFound />} />
       </Routes>
     </Suspense>
