+++ packages/pn-personafisica-webapp/src/pages/PrivacyPolicy.page.tsx
@@ -1,0 +1,403 @@
+import {
+  Box,
+  Container,
+  Link,
+  Typography
+} from "@mui/material";
+
+const PrivacyPolicyPage = () => (
+    <>
+      <Container
+        maxWidth="xl"
+        sx={{py: 4}}
+        >
+        <Typography variant="h2">Informativa sul trattamento dei dati personali</Typography>
+        <Typography variant="h4">Piattaforma Notifiche Digitali</Typography>
+        <Typography variant="h5">Destinatari persone fisiche</Typography>
+        <Box mt={3}>
+          <Typography variant="body1">
+            La presente informativa, resa ai sensi degli artt. 13-14 del Regolamento (UE) 2016/679 (di seguito
+            il <b>“Regolamento”</b> o <b>“GDPR”</b>), descrive le modalità del trattamento dei dati personali degli
+            utenti destinatari delle notifiche, dei soggetti da essi delegati e delle persone di loro fiducia (di
+            seguito <b>“Interessato/i”</b> o genericamente <b>“Utente/i”</b>) che aderiscono alla piattaforma
+            tramite la quale le pubbliche amministrazioni aderenti (<b>“Pubbliche Amministrazioni”</b>) notificano
+            atti, avvisi e comunicazioni (di seguito <b>“Piattaforma”</b>) gestita da PagoPA S.p.A. (di seguito
+            anche <b>“Titolare”</b> o <b>“Società”</b>) ai sensi dell’art. 26 del decreto-legge n. 76 del 17 luglio
+            2020 come convertito, con modificazioni, dalla legge 11 settembre 2020, n. 120 e come altresì modificato
+            dal decreto-legge. n. 77 del 31 maggio 2021, così come convertito dalla legge 29 luglio 2021 n. 108, e
+            raggiungibile all’indirizzo web: <Link href="https://pn.pagopa.it">https://pn.pagopa.it</Link>. (di
+            seguito il <b>“Sito”</b>).
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            La validità dell’informativa contenuta nella presente pagina è limitata alla sola Piattaforma e non si
+            estende ad altri siti web eventualmente consultabili mediante collegamento ipertestuale.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            <u>Le informative sul trattamento dei dati personali effettuati dalle Pubbliche Amministrazioni mittenti
+              possono essere rinvenute ai seguenti link:</u><br/>
+            <Link href="https://selfcare.notifichedigitali.it/privacy-tos">Comune di Milano</Link>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            Titolare del trattamento dei dati e base giuridica
+          </Typography>
+          <Typography variant="body1">
+            Il Titolare del trattamento è PagoPA S.p.A., con sede in Roma, Piazza Colonna 370, CAP - 00187, n. di
+            iscrizione a Registro Imprese di Roma, CF e P.IVA 15376371009.<br/>
+            E-mail: <Link href="mailto:info@pagopa.it">info@pagopa.it.</Link><br/>
+            La Società tratta i dati personali in qualità di Titolare del trattamento sulla base dell’esercizio del
+            compito di interesse pubblico di cui all’art. 6, comma 1, lett. e), del GDPR riguardante, in
+            particolare, la progettazione, la gestione e lo sviluppo della Piattaforma.<br/>
+            Per ogni domanda inerente il trattamento di dati personali si prega di scrivere utilizzando il <Link
+            href="https://privacyportal-de.onetrust.com/webform/77f17844-04c3-4969-a11d-462ee77acbe1/9ab6533d-be4a-482e-929a-0d8d2ab29df8">form</Link> dedicato
+            alla gestione delle richieste degli interessati.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            Data protection officer (DPO) o Responsabile Protezione Dati (RPD)
+          </Typography>
+          <Typography variant="body1">
+            PagoPA S.p.A. ha nominato un proprio Responsabile della Protezione dei dati, ai sensi dell’art. 37 del
+            Regolamento, che può essere contattato tramite il presente <Link
+            href="https://privacyportal-de.onetrust.com/webform/77f17844-04c3-4969-a11d-462ee77acbe1/9ab6533d-be4a-482e-929a-0d8d2ab29df8">form</Link> di
+            contatto. Il form è altresì disponibile sul sito web della Società nella sezione “Diritto alla
+            protezione dei dati personali”.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            Categorie di soggetti ai quali i dati personali possono essere comunicati e finalità della comunicazione
+          </Typography>
+          <Typography variant="body1" component="span">
+            La Società potrà comunicare alcuni dati personali a soggetti terzi dei quali si avvale per lo
+            svolgimento di attività connesse alla gestione della Piattaforma.<br/>
+            In particolare, i dati personali potranno essere comunicati a società esterne che offrono alla Società
+            taluni servizi e ai soggetti cui la comunicazione sia dovuta in forza di obblighi di legge. I suddetti
+            soggetti potranno trattare i dati in qualità di responsabili per conto della Società o di titolari
+            autonomi nel rispetto delle disposizioni di legge.<br/>
+            L’elenco completo dei responsabili del trattamento può essere richiesto alla Società scrivendo sul <Link
+            href="https://privacyportal-de.onetrust.com/webform/77f17844-04c3-4969-a11d-462ee77acbe1/9ab6533d-be4a-482e-929a-0d8d2ab29df8">form</Link> di
+            contatto presente nella sezione “Diritto alla protezione dei dati personali” del <Link
+            href="https://www.pagopa.it/it/">sito web</Link> della Società.<br/>
+            Non è prevista alcuna forma di diffusione dei dati personali a soggetti indeterminati.<br/>
+            Alcuni dei responsabili del trattamento di cui la Società si avvale vengono elencati, a titolo
+            esemplificativo, di seguito:<br/>
+            <ul>
+              <li><u>Fornitore del servizio universale (di seguito “FSU”) ovvero Poste Italiane S.p.A.</u>, con sede
+                legale in Roma, Viale Europa n. 190, Codice Fiscale e numero di iscrizione al Registro delle Imprese
+                di Roma 97103880585/1996, P.IVA 01114601006. Il DPO è reperibile presso l’ufficio del Responsabile
+                della Protezione dei Dati di Poste Italiane, in viale Europa, 175 - 00144 Roma,
+                ufficiorpd@posteitaliane.it;
+              </li>
+              <li><u>Amazon Web Services EMEA SARL</u> - indirizzo: 38 Avenue John F. Kennedy, L-1855 Lussemburgo.
+                Il responsabile per la protezione dei dati può essere contattato al seguente indirizzo:
+                aws-EU-privacy@amazon.com;
+              </li>
+              <li><u>OneTrust Technology Limited</u> - per la gestione delle richieste degli interessati, con sede
+                in 82 St John Street, London, England, EC1M 4JN, n. di iscrizione al Registro Imprese Great England
+                and Wales n. 04156317. Gli interessati possono contattare il responsabile della protezione dei dati
+                scrivendo all’indirizzo e-mail: <Link href="mailto:dpo@onetrust.com">dpo@onetrust.com.</Link></li>
+            </ul>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            Dati personali trattati e finalità del trattamento
+          </Typography>
+          <Typography variant="body1" component="span">
+            <u>1. Ai fini dell’accesso e gestione delle preferenze di Piattaforma da parte delle persone
+              fisiche.</u><br/>
+            <u>1.1 La Società, al fine di identificare univocamente ogni Utente che accede alla Piattaforma, tratta
+              i seguenti dati:</u>
+            <ul>
+              <li>dati acquisiti a mezzo SPID/SPID Professionale o CIE quali nome, cognome e codice fiscale;</li>
+              <li>token identificativo e ID utente generato all’accesso alla Piattaforma con SPID/SPID
+                Professionale/CIE;
+              </li>
+            </ul>
+            <u>1.2 Con riferimento invece alla gestione delle preferenze, e dunque anche al conferimento delle
+              deleghe, la Società tratta i seguenti dati: </u>
+            <ul>
+              <li>Dati acquisiti a mezzo SPID/SPID Professionale o CIE quali nome, cognome e codice fiscale del
+                destinatario e dell’eventuale delegato;
+              </li>
+              <li>rimozione o modifica dei delegati e relativi dati di quest’ultimo quali: nome, cognome, codice
+                fiscale, recapito digitale;
+              </li>
+              <li>data di scadenza della delega;</li>
+              <li>recapiti digitali e codici di conferma a 5 cifre per la validazione del recapito digitale;</li>
+              <li>codice di accettazione della delega trasmesso al delegato al destinatario con mezzi e modalità che
+                desidera;
+              </li>
+              <li>domicilio generale di piattaforma.</li>
+            </ul>
+          </Typography>
+          <Typography variant="body1" component="span">
+            <u>2. Per l’invio della notifica digitale la Società tratta i seguenti dati del destinatario:</u>
+            <ul>
+              <li>codice fiscale per l’interrogazione sui registri nazionali delle PEC ai fini dell’acquisizione del
+                domicilio digitale generale;
+              </li>
+              <li>domicilio di piattaforma se il destinatario ha eletto tale domicilio all’atto della autenticazione
+                in Piattaforma come primo domicilio di invio;
+              </li>
+              <li>domicilio speciale se la Pubblica Amministrazione lo ha comunicato alla Società all’atto della
+                trasmissione degli atti oggetto di notifica;
+              </li>
+              <li>atti oggetto di notifica e documenti relativi al pagamento e avviso di avvenuta ricezione
+                (contenente a titolo esemplificativo i seguenti dati: pubblica amministrazione mittente, data e ora
+                di messa a disposizione dell’atto in Piattaforma; nome, cognome e codice fiscale; oggetto della
+                notifica, ovvero causale/titolo dell’avviso/atto/provvedimento/comunicazione; codice IUN; relativo
+                atto opponibile ai terzi generato dall’invio della notifica.
+              </li>
+            </ul>
+          </Typography>
+          <Typography variant="body1" component="span">
+            <u>3. Nel caso in cui un destinatario non possa accedere alla Piattaforma per propria scelta o perché
+              non in possesso di identità digitale, la Società mette a disposizione dello stesso una modalità
+              diversificata di accesso alla Piattaforma che prevede che l’Utente possa recarsi presso il FSU e
+              procedere alla configurazione di tale accesso semplificato (richiesta OTP). A tal fine la Società
+              tratta i seguenti dati:</u>
+            <ul>
+              <li>recapito digitale, ovvero e-mail, cellulare e app IO, associato al codice fiscale del
+                destinatario;
+              </li>
+              <li>estremi del documento di riconoscimento degli Utenti;</li>
+              <li>delega e relativa firma in esso contenuta, nonché relativa copia conforme;</li>
+              <li>indirizzo fisico degli Utenti;</li>
+              <li>nome, cognome e codice fiscale del destinatario;</li>
+              <li>OTP, ovvero una password di identificazione temporanea trasmessa su uno dei recapiti digitali
+                censiti per l’abilitazione all’accesso semplificato.
+              </li>
+            </ul>
+            3.1 Tale accesso semplificato, protetto e temporaneo è altresì disponibile all’Utente che si identifica
+            attraverso <u>SPID/SPID Professionale o CIE (e dunque senza necessità di OTP)</u>. A tal fine la Società
+            tratta, oltre a quanto sopra e ad esclusione dell’OTP, anche i seguenti dati:
+            <ul>
+              <li>dati acquisiti a mezzo SPID/SPID Professionale o CIE quali nome, cognome, codice fiscale degli
+                Utenti.
+              </li>
+            </ul>
+            <strong>Nel caso in cui Tu stia inserendo dati personali di terzi, ti invitiamo ad ottenere specifica
+              autorizzazione a procedere da parte di tale soggetto, al fine di non commettere un trattamento
+              illecito.</strong>
+            <br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1">
+            <u>4. Con riferimento alla raccolta dei dati relativi alla gestione delle preferenze di navigazione per
+              l’ottimizzazione della Piattaforma, e al tracciamento dei log di sistema/funzionamento e di accesso
+              per garantire la sicurezza della Piattaforma</u>, la Società sui i sistemi informatici preposti al
+            funzionamento della Piattaforma acquisisce, nel corso del loro normale esercizio, alcuni dati personali
+            la cui trasmissione è implicita nell’uso dei protocolli di comunicazione di Internet. In questa
+            categoria di dati rientrano gli indirizzi IP o i nomi a dominio dispositivi utilizzati dagli Utenti, gli
+            indirizzi in notazione URI/URL (Uniform Resource Identifier/Locator) delle risorse richieste, l’orario
+            della richiesta, il tipo di richiesta effettuata (risorsa richiesta e nome dell’operazione), il metodo
+            utilizzato nel sottoporre la richiesta al server, la dimensione del file ottenuto in risposta, il codice
+            numerico indicante lo stato della risposta data dal server (buon fine, errore, ecc.) ed altri parametri
+            relativi al sistema operativo e all’ambiente informatico dell’Utente. A ciascun Utente, i sistemi della
+            Piattaforma attribuiscono un codice identificativo univoco associato all’account e conservato nei log,
+            nonché un token di sessione. Tali dati, necessari per la fruizione delle funzionalità della Piattaforma,
+            vengono anche trattati allo scopo di controllare il corretto funzionamento dei sistemi e dei servizi
+            offerti dalla Piattaforma (diagnostica), nonché per motivi di sicurezza.
+          </Typography>
+          <br/>
+          <Typography variant="body1">
+            <strong>
+              Inoltre, la totalità dei dati personali trattati possono essere utilizzati dal Titolare del
+              trattamento in giudizio o nelle fasi propedeutiche alla sua eventuale instaurazione per la difesa da
+              abusi perpetrati dall’Utente.
+            </strong>
+          </Typography>
+          <Typography variant="body1">
+            I dati personali degli Utenti sono trattati adottando adeguate misure di sicurezza volte ad impedire
+            l’accesso, la divulgazione, la modifica o la distruzione non autorizzata dei dati personali. Tutti i
+            dipendenti della Società che hanno accesso ai dati personali sono debitamente designati quali soggetti
+            autorizzati al trattamento e ciascuno di essi è incaricato di trattare unicamente i dati strettamente
+            necessari allo svolgimento delle proprie mansioni lavorative. Il trattamento è effettuato
+            prevalentemente mediante strumenti informatici, con modalità organizzative e logiche strettamente
+            correlate alle finalità sopra indicate. Oltre al Titolare, in alcuni casi, possono avere accesso ai
+            dati, a seguito di comunicazione da parte dello stesso, anche ulteriori soggetti coinvolti nella
+            gestione e l’erogazione dei servizi offerti tramite la Piattaforma, nominati, ove necessario, quali
+            responsabili del trattamento ai sensi dell’art. 28 del Regolamento. Tali soggetti trattano i dati
+            esclusivamente per attività funzionali o comunque strettamente connesse allo svolgimento del Servizio.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            Categorie di destinatari dei dati
+          </Typography>
+          <Typography variant="body1" component="span">
+            Il Titolare, nello svolgimento delle proprie attività, ha la facoltà di trasmettere, eventualmente e su
+            richiesta, i dati alle seguenti categorie di destinatari:
+            <ul>
+              <li>responsabili del trattamento quali fornitori di servizi;</li>
+              <li>altri soggetti pubblici o privati che ne facciano richiesta per l’esecuzione di un compito di
+                interesse pubblico o connesso all’esercizio di un pubblico potere o per adempiere a un obbligo
+                legale o contrattuale.
+              </li>
+            </ul>
+          </Typography>
+          <Typography variant="body1" fontWeight="bold">
+            Tempi di conservazione dei dati
+          </Typography>
+          <Typography variant="body1">
+            I dati personali relativi agli Utenti sono trattati per il tempo strettamente necessario al
+            perseguimento delle finalità per le quali sono stati raccolti. Nel rispetto dei principi di
+            proporzionalità e necessità, i dati non sono conservati per periodi più lunghi rispetto a quelli
+            indispensabili alla realizzazione delle finalità sopra indicate e dunque al diligente svolgimento delle
+            attività.
+          </Typography>
+          <br/>
+          <Typography variant="body1" component="span">
+            Di seguito alcuni dei tempi di conservazione stabiliti:
+            <ul>
+              <li>i log di accesso a mezzo SPID/SPID Professionale e CIE: 2 anni;</li>
+              <li>i log di funzionamento ovvero log di sistema: 3 anni;</li>
+              <li>gli audit log diversi da quelli specifici sugli atti opponibili ai terzi: 5 anni;</li>
+              <li>i dati di navigazione ovvero i record contenenti informazioni relativi a browser, IP e device
+                utilizzati durante le interazioni dell’Utente sulla Piattaforma: 90 giorni (nello specifico anche IP
+                di navigazione estratto dal log);
+              </li>
+              <li>il codice OTP (a 5 cifre): 1 ora;</li>
+              <li>codice di accettazione del recapito digitale (a 5 cifre): 1 ora.</li>
+            </ul>
+          </Typography>
+          <Typography variant="body1" component="span">
+            <ul>
+              <li>gli audit log specifici sugli atti opponibili ai terzi, le informazioni necessarie alla creazione
+                dei QR code e link HTML, a partire dalla data di perfezionamento della notifica;
+              </li>
+              <li>atto di delega scansionato/copia conforme per attivazione OTP (per 10 anni dalla richiesta).</li>
+            </ul>
+            Per maggiori informazioni è sempre disponibile il <Link
+            href="https://privacyportal-de.onetrust.com/webform/77f17844-04c3-4969-a11d-462ee77acbe1/9ab6533d-be4a-482e-929a-0d8d2ab29df8">form</Link> dedicato
+            alla gestione delle richieste degli interessati disponibile nella sezione “Diritto alla protezione dei
+            dati personali” del sito internet della Società <Link
+            href="https://www.pagopa.it/it/">https://www.pagopa.it/it/</Link>.
+          </Typography>
+          <Typography variant="body1" fontWeight="bold">
+            Trasferimento transfrontaliero dei dati
+          </Typography>
+          <Typography variant="body1">
+            I dati personali potranno essere trasferiti fuori dall’Unione Europea da parte di fornitori di servizi
+            di cui la Società si avvale per attività connesse alla gestione della Piattaforma. Tale trasferimento,
+            ove ricorra il caso, verrà disciplinato con i fornitori di servizi mediante il ricorso a clausole
+            contrattuali standard adottate dalla Commissione europea con la decisione 2021/914/UE ed eventuali
+            successive modifiche o, in alternativa, sulla base di una decisione di adeguatezza della Commissione e/o
+            di ogni altro strumento consentito dalla normativa di riferimento. La lista dei fornitori coinvolti,
+            comprensivi delle garanzie poste in essere, può essere richiesta in qualsiasi momento, utilizzando
+            il <Link
+            href="https://privacyportal-de.onetrust.com/webform/77f17844-04c3-4969-a11d-462ee77acbe1/9ab6533d-be4a-482e-929a-0d8d2ab29df8">form</Link> dedicato
+            alla gestione delle richieste degli interessati disponibile nella sezione “Diritto alla protezione dei
+            dati personali” del sito internet della Società <Link
+            href="https://www.pagopa.it/it/">https://www.pagopa.it/it/</Link>.
+          </Typography>
+          <Typography variant="body1" fontWeight="bold">
+            Diritti degli interessati
+          </Typography>
+          <Typography variant="body1">
+            Gli Utenti, ai quali i dati personali si riferiscono, hanno il diritto di ottenere dalla Società, in
+            qualità di Titolare del trattamento, l’accesso ai propri dati personali, l’aggiornamento,
+            l’integrazione, la rettifica o, laddove previsto dalla legge e nei limiti previsti, la cancellazione
+            degli stessi, la limitazione del trattamento e il diritto di opporsi allo stesso. Gli Utenti potranno,
+            altresì, chiedere l’anonimizzazione dei dati o il blocco dei dati trattati in violazione di legge,
+            compresi quelli di cui non è necessaria la conservazione in relazione agli scopi del trattamento.
+            <br/>
+            Le richieste dovranno essere inoltrate alla Società. Gli Interessati possono, altresì, contattare il
+            Responsabile della protezione dei dati per tutte le questioni inerenti il trattamento dei propri dati
+            personali e l’esercizio dei propri diritti, utilizzando il presente <Link
+            href="https://privacyportal-de.onetrust.com/webform/77f17844-04c3-4969-a11d-462ee77acbe1/9ab6533d-be4a-482e-929a-0d8d2ab29df8">form</Link> dedicato
+            alla gestione delle richieste degli interessati disponibile nella sezione “Diritto alla protezione dei
+            dati personali” del sito internet della Società <Link
+            href="https://www.pagopa.it/it/">https://www.pagopa.it/it/</Link>.
+          </Typography>
+          <br/>
+          <Typography variant="body1">
+            Gli Interessati che ritengano che il trattamento dei dati personali a loro riferiti, effettuato dal
+            Titolare del trattamento, avvenga in violazione di quanto previsto dal Regolamento, hanno il diritto di
+            proporre reclamo al Garante, come previsto dall’art. 77 del Regolamento stesso, o di adire le opportune
+            sedi giudiziarie (art. 79 del Regolamento).
+            <br/>
+            Di seguito si trasmettono i dati di contatto ed i riferimenti dell’<b>Autorità di controllo - Garante
+            per la protezione dei dati personali</b>
+            <br/>
+            Indirizzo e-mail: garante@gpdp.it
+            <br/>
+            Indirizzo PEC: protocollo@pec.gpdp.it
+            <br/>
+            Sito web: https://www.garanteprivacy.it
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">
+            Modifiche
+          </Typography>
+          <Typography variant="body1">
+            Il Titolare si riserva il diritto di apportare alla presente Informativa, a propria esclusiva
+            discrezione ed in qualunque momento, tutte le modifiche ritenute opportune o rese obbligatorie dalle
+            norme di volta in volta vigenti, dandone adeguata pubblicità agli Utenti. In caso di mancata
+            accettazione delle modifiche effettuate, si invitano gli Utenti a cessare l’utilizzo della Piattaforma e
+            a chiedere alla Società la rimozione dei propri dati personali; salvo quanto diversamente specificato,
+            la precedente versione dell’informativa continuerà ad applicarsi ai dati personali raccolti sino a quel
+            momento.
+          </Typography>
+          <Typography variant="body1" fontWeight="bold" textAlign="center">
+            Cookie policy
+          </Typography>
+          <Typography variant="body1" component="span">
+            Il presente Avviso Cookie è parte della nostra Informativa sulla privacy. Per ulteriori informazioni su
+            di noi, e su come proteggiamo informazioni degli Utenti, si prega di consultare la nostra Informativa
+            sulla privacy.
+            <br/><br/>
+            Il Sito utilizza cookie tecnici di navigazione o di sessione che garantiscono la normale navigazione e
+            fruizione del Sito (essi vengono memorizzati sul terminale dell’Utente sino alla chiusura del browser).
+            Detti cookie sono utilizzati nella misura strettamente necessaria per rendere il servizio relativo alla
+            Piattaforma. Essi garantiscono un’adeguata fruizione del Sito e consentono all’Utente di navigare e
+            utilizzare servizi e opzioni. Il loro utilizzo esula da scopi ulteriori e tali cookie sono installati
+            direttamente dal Titolare (rientrano, dunque, nella categoria di cookie di prima parte o proprietari)
+            <br/><br/>
+            Per l’installazione dei suddetti cookie non è richiesto il preventivo consenso degli Utenti, mentre
+            resta fermo l’obbligo di dare l’informativa ai sensi dell’art. 13 del Regolamento e dell’art. 122 del D.
+            Lgs. 196/2003 e ss.mm.ii. (“Codice Privacy”).
+            <br/><br/>
+            Il Sito presenta cookie di terze parti, creati e gestiti da soggetti diversi dal Titolare, e sui quali
+            lo stesso non ha e non può avere alcun controllo. Nello specifico, l’unico cookie settato di terze parti
+            è MixPanel utile per il tracciamento degli eventi di failure nelle operazioni, di cui viene richiesto
+            preventivo consenso.
+            <br/><br/>
+            I cookie vengono conservati per un periodo non superiore a 6 mesi dal momento della raccolta, salvo che
+            la loro ulteriore conservazione non si renda necessaria per l’accertamento di reati.
+            <br/><br/>
+            &Egrave; possibile gestire le preferenze relative ai cookie attraverso le opzioni fornite dal proprio
+            browser e dal cookie banner nella sezione scopri di più dove potrai accettare o meno tutti o alcuni dei
+            cookie presenti.
+            <br/><br/>
+            Di seguito invece le istruzioni rese disponibili dai relativi fornitori ai link di seguito indicati:
+            <ul>
+              <li><Link
+                href="https://support.google.com/chrome/answer/95647?co=GENIE.Platform=Desktop&hl=it">Chrome</Link>
+              </li>
+              <li><Link
+                href="https://support.mozilla.org/it/kb/Attivare%20e%20disattivare%20i%20cookie">Firefox</Link></li>
+              <li><Link href="https://support.apple.com/kb/ph19214?locale=it_IT">Safari</Link></li>
+              <li><Link
+                href="https://support.microsoft.com/it-it/help/17442/windows-internet-explorer-delete-manage-cookies">Internet
+                Explorer</Link></li>
+              <li><Link href="https://support.microsoft.com/it-it/help/4027947/windows-delete-cookies">Edge</Link>
+              </li>
+              <li><Link href="https://help.opera.com/en/latest/web-preferences/#cookies">Opera</Link></li>
+            </ul>
+            <Typography variant="body1">Data di ultimo aggiornamento: 20.07.2022</Typography>
+            <br/>
+            <Typography variant="body1">
+              I Tuoi feedback sono importanti per migliorare il Sito, anche sulle questioni privacy! Se qualcosa che
+              riguarda la Tua privacy non ti convince, non Ti è chiaro o ritieni che dovremmo agire diversamente,
+              scrivici utilizzando il <Link
+              href="https://privacyportal-de.onetrust.com/webform/77f17844-04c3-4969-a11d-462ee77acbe1/9ab6533d-be4a-482e-929a-0d8d2ab29df8">form</Link> dedicato
+              alla gestione delle richieste degli interessati.
+            </Typography>
+          </Typography>
+        </Box>
+      </Container>
+    </>);
+
+export default PrivacyPolicyPage;
