+++ packages/pn-personafisica-webapp/src/pages/TermsOfService.page.tsx
@@ -1,52 +1,364 @@
-import { Fragment, ReactNode, useEffect } from 'react';
-import { Trans } from 'react-i18next';
-import { useNavigate } from 'react-router-dom';
-import { Grid } from '@mui/material';
-import { TermsOfServiceHandler, useIsMobile } from '@pagopa-pn/pn-commons';
+import {
+  Box,
+  Container,
+  Link,
+  Typography
+} from "@mui/material";
 
-import { useAppDispatch, useAppSelector } from '../redux/hooks';
-import { acceptToS } from '../redux/auth/actions';
-import * as routes from '../navigation/routes.const';
-import { RootState } from '../redux/store';
+const PrivacyPolicyPage = () => (
+    <>
+      <Container
+        maxWidth="xl"
+        sx={{py: 4}}
+        >
+        <Typography variant="h2">Termini e condizioni d’uso</Typography>
 
-const TermsOfService = () => {
-  const isMobile = useIsMobile();
-  const dispatch = useAppDispatch();
-  const navigate = useNavigate();
-  const tos = useAppSelector((state: RootState) => state.userState.tos);
+        <Typography variant="h4">Piattaforma Notifiche Digitali</Typography>
+        <Typography variant="h5">Destinatari persone fisiche</Typography>
+        <Box mt={3}>
+          <Typography variant="body1" fontWeight="bold">Data ultimo aggiornamento: 08/08/2022</Typography>
+          <br/>
+          <Typography variant="body1">
+            I presenti termini e condizioni d’uso (<b>“Termini & Condizioni d’uso” o “ToS”</b>) regolano il rapporto
+            tra l’<b>utente persona fisica</b> (di seguito <b>“Utente” o “Tu”</b>) e PagoPA S.p.A. (di
+            seguito <b>“Società”</b>), quale società per legge avente la gestione della Piattaforma (come nel
+            seguito definita), stabilendo diritti e obblighi connessi al suo utilizzo.<br/>
+            <b>I presenti ToS regolano quindi il rapporto tra gli Utenti del sito <Link
+              href="https://cittadini.notifichedigitali.it">https://cittadini.notifichedigitali.it</Link></b> (di
+            seguito <b>“Sito”</b>) che accedono alla Piattaforma Notifiche Digitali (di seguito <b>“Piattaforma”</b>)
+            di cui ai seguenti riferimenti normativi:
+          </Typography>
+          <ul>
+            <li>
+              <Typography variant="body1" sx={{display: "list-item"}}>art. 1, comma 402, della legge 27 dicembre
+                2019, n. 160;</Typography>
+            </li>
+            <li>
+              <Typography variant="body1">art. 26 del decreto-legge del 17 luglio 2020, n. 76 come convertito, con
+                modificazioni, dalla legge 11 settembre 2020, n. 120 e come altresì modificato dal decreto-legge del
+                31 maggio 2021, n. 77 così come convertito dalla legge 29 luglio 2021 n. 108 (di seguito
+                “Decreto”);</Typography>
+            </li>
+            <li>
+              <Typography variant="body1">decreto del Ministro per la Trasformazione Digitale dell’8 febbraio 2022,
+                n. 58 dal titolo “Regolamento recante piattaforma per la notificazione degli atti della pubblica
+                amministrazione” (di seguito “Decreto Funzionamento”);</Typography>
+            </li>
+            <li>
+              <Typography variant="body1">decreto del Ministro per la Trasformazione Digitale del 30 maggio 2022 dal
+                titolo “Individuazione dei costi e dei criteri e modalità di ripartizione e ripetizione delle spese
+                di notifica degli atti tramite la piattaforma di cui all’art. 26, comma 14 del decreto-legge 16
+                luglio 2020, n. 76” (di seguito “Decreto Costi”).</Typography>
+            </li>
+          </ul>
+          <Typography variant="body1">
+            <strong>
+              Leggili attentamente: utilizzando il Sito, li accetti integralmente e ti impegni a rispettarli.
+            </strong>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">1. Descrizione del servizio </Typography>
+          <Typography variant="body1" component="span">
+            1.1 Tramite la Piattaforma, le pubbliche amministrazioni mittenti (di seguito <b>“Pubbliche
+            Amministrazioni”</b>) notificano atti, provvedimenti, avvisi e comunicazioni (di seguito
+            genericamente <b>“documenti informatici”</b>) agli Utenti, ossia ai loro destinatari, i quali possono
+            servirsi anche di delegati o altre persone di fiducia per accedere agli stessi documenti.<br/>
+            1.2 Per le notifiche effettuate tramite Piattaforma, Tu potresti non ricevere materialmente il documento
+            informatico oggetto di notifica ma certamente riceverai un avviso (avviso di avvenuta ricezione o avviso
+            di mancato recapito) che ti informerà della notifica eseguita nei Tuoi confronti e che ti permetterà,
+            altresì, di reperire i documenti informatici di riferimento della notifica, come disponibili sulla
+            Piattaforma.<br/>
+            1.3 In coerenza con i principi generali del diritto, la notifica dei documenti informatici per il
+            mittente e per i destinatario si perfeziona in due momenti diversi e distinti, e precisamente:<br/>
+            <ul>
+              <li>per le Pubbliche Amministrazioni (mittenti) la notifica si perfeziona con la messa a disposizione
+                dei medesimi sulla Piattaforma;
+              </li>
+              <li>per gli Utenti (destinatari) invece la notifica si perfeziona:<br/>
+                <b>1a)</b> il <b>settimo giorno</b> successivo alla data di consegna dell’<b>avviso di avvenuta
+                  ricezione</b> se in formato <b>elettronico</b>;<br/>
+                <b>1b)</b> il <b>quindicesimo giorno</b> successivo alla data del deposito dell’<b>avviso di mancato
+                  recapito</b>;<br/>
+                <b>2)</b> il <b>decimo</b> giorno successivo al perfezionamento della notificazione dell’avviso di
+                avvenuta ricezione se in formato cartaceo;<br/>
+                <b>3)</b> in ogni caso, se anteriore, <b>nella data</b> in cui Tu o un Tuo delegato, <b>hai/ha
+                  accesso alla Tua area riservata</b> della Piattaforma, contenente il dettaglio della notifica.
+              </li>
+            </ul>
+          </Typography>
+          <Typography variant="body1">
+            1.4 Potrai per il tramite della Piattaforma, per ogni documento informatico oggetto di notifica
+            visualizzarne la Pubblica Amministrazione mittente, la data e l’ora di messa a disposizione dello stesso
+            in Piattaforma, lo storico del processo di notificazione, ivi inclusi i relativi atti opponibili ai
+            terzi, gli avvisi di mancato recapito e il codice identificativo univoco della notifica (IUN).<br/>
+            1.5 Ai sensi dell’articolo 12 del Decreto Funzionamento, infatti, la Piattaforma genera degli atti
+            opponibili ai terzi (anche chiamate <b>“Attestazioni”</b>), ossia degli atti giuridici idonei ad
+            esprimere la loro efficacia anche nei confronti dei terzi.<br/>
+            1.6 La messa a disposizione dei documenti informatici oggetto di notifica in Piattaforma impedisce
+            qualsiasi decadenza e interrompe il termine di prescrizione correlato alla notificazione dei documenti
+            informatici oggetto di notifica.<br/>
+            1.7 Accettando i presenti ToS, Tu dichiari espressamente di volerti avvalere della Piattaforma e dei
+            servizi di notifica offerti per il tramite di questa (i <b>“Servizi”</b>), alle condizioni e nei termini
+            di seguito descritti.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">2. Identificazione e accesso alla Piattaforma</Typography>
+          <Typography variant="body1">
+            2.1 Affinché Tu possa accedere alla Piattaforma e usufruire digitalmente dei Servizi, dovrai
+            identificarti tramite sistema pubblico per la gestione dell’identità digitale di cittadini e imprese
+            (SPID) ovvero tramite la Carta d’identità elettronica (CIE).<br/>
+            2.2 Dopo aver preso visione dei presenti ToS, avrai diritto ad accedere alla Tua area riservata della
+            Piattaforma, per il reperimento, la consultazione e l’acquisizione dei documenti informatici oggetto di
+            notifica (vedi anche 9.1 viii).<br/>
+            2.3 Nei casi in cui Tu non possa utilizzare o non sia in possesso di SPID o CIE, hai comunque a
+            disposizione una modalità alternativa di accesso, presso Poste Italiane S.p.A., quale Fornitore del
+            Servizio Universale (di seguito <b>“FSU”</b>) che ti consente di acquisire l’estrazione della copia
+            analogica dei documenti informatici oggetto di notifica, degli atti opponibili ai terzi rilasciati dalla
+            Società ai sensi del successivo art. 9 e dei documenti di pagamento, nonché di accedere agli avvisi di
+            avvenuta ricezione non consegnati.<br/>
+            Tale modalità tramite Poste Italiane S.p.A. necessita, oltre che di avere l’avviso di avvenuta
+            ricezione, anche di identificare di persona il destinatario a mezzo di un documento di riconoscimento e
+            del codice fiscale oppure, in caso di delega, verrà identificato di persona il delegato per cui si
+            aggiungono ai documenti richiesti per il destinatario anche gli stessi per il delegato.<br/>
+            2.4 E’ possibile acquisire la copia analogica dei documenti di cui al punto 2.3 sia per il destinatario,
+            sia per il delegato, dotati di SPID o CIE, direttamente allo sportello di FSU.
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">3. Delega per l’accesso - autodichiarazione della persona di
+            fiducia</Typography>
+          <Typography variant="body1">
+            3.1 I destinatari, possono conferire, tramite delega, a soggetti terzi il potere di accedere alla
+            Piattaforma per reperire, consultare e acquisire, per loro conto, i documenti informatici oggetto di
+            notifica.<br/>
+            3.2 La delega conferita contiene nome, cognome e codice fiscale del delegato, e può essere conferita per
+            tutte o anche soltanto per alcune specifiche Pubbliche Amministrazioni. A tal fine, la Piattaforma
+            elabora un codice di accettazione della delega da comunicare al delegato, che viene da quest’ultimo
+            utilizzato per accettare la stessa.<br/>
+            3.3 Tu puoi revocare la delega già conferita in qualunque momento tramite la Piattaforma, così come il
+            delegato sempre tramite Piattaforma ha la facoltà di rinunciare in ogni tempo alla delega ricevuta.<br/>
+            3.4 La Piattaforma mette ogni mese a disposizione un promemoria delle sole deleghe attive e consente al
+            delegante di monitorare ogni accesso effettuato dal delegato.<br/>
+            3.5 E’ inoltre per Te possibile far accedere presso l’FSU una persona di fiducia ai Tuoi documenti
+            informatici oggetto di notifica. A tal fine la persona di fiducia dovrà presentare un’autodichiarazione
+            ex art. 46 del DPR 28.12.2000, n. 445, attestante l’incarico da Te ricevuto.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">4. Elezione domicilio digitale e invio digitale</Typography>
+          <Typography variant="body1">
+            4.1 Accedendo alla Piattaforma i documenti informatici oggetto di notifica potranno essere inoltrati al
+            Tuo domicilio digitale, quale quello di piattaforma (se eletto direttamente presso la Piattaforma),
+            quello speciale (se eletto direttamente presso la Pubblica Amministrazione) o quello generale (se
+            presente in un elenco pubblico previsto dalla legge).<br/>
+            4.2 Qualora Tu non abbia mai eletto un domicilio digitale e Tu non sia neppure un soggetto iscritto ad
+            un albo professionale (es. avvocato, commercialista, ecc.) o comunque iscritto ad un registro nazionale
+            in cui la Tua PEC è resa pubblica, puoi eleggere un domicilio digitale di Piattaforma; in tal caso,
+            accedi alla Piattaforma e indica il corrispondente indirizzo, che verrà utilizzato dalla Società quale
+            gestore della Piattaforma per il recapito dei tuoi documenti informatici oggetto di notifica da parte di
+            ogni Pubblica Amministrazione tempo per tempo utilizzatrice della Piattaforma, senza eccezione alcuna in
+            coerenza con quanto indicato al successivo punto 4.6 e salvo quanto previsto al successivo punto
+            4.5.<br/>
+            4.3 Qualora Tu non abbia eletto presso la Pubblica Amministrazione un domicilio digitale speciale, ma Tu
+            sia invece un soggetto iscritto ad un albo professionale (es. avvocato, commercialista, ecc.) o comunque
+            iscritto ad un registro nazionale in cui la PEC è resa pubblica, la Piattaforma - attraverso il Tuo
+            codice fiscale - verifica l’esistenza di tale indirizzo PEC e lo utilizza in autonomia per inviarti i
+            documenti informatici oggetto di notifica.<br/>
+            4.4 Se avendo Tu un domicilio digitale speciale e/o un domicilio generale preferisci che le Pubbliche
+            Amministrazioni utilizzatrici della Piattaforma usino un Tuo diverso domicilio digitale puoi eleggere un
+            domicilio digitale di Piattaforma; in tal caso, accedi alla Piattaforma e indica il corrispondente
+            indirizzo, che verrà utilizzato dalla Piattaforma in via prioritaria per il recapito dei Tuoi documenti
+            informatici oggetto di notifica.<br/>
+            4.5 Ricorda che Tu hai anche la facoltà di indicare dei domicili digitali di Piattaforma diversificati
+            in relazione alle varie Pubbliche Amministrazioni (mittenti), ma senza distinzione o limitazione alcuna
+            in base a specifici atti e/o affari.<br/>
+            4.6 Nel caso in cui Tu abbia indicato un solo domicilio digitale di Piattaforma, tale indirizzo verrà
+            utilizzato per l’invio dei Tuoi documenti informatici oggetto di notifica da parte di tutte le Pubbliche
+            Amministrazioni, sempre senza distinzione o limitazione alcuna in base a specifici atti e/o affari.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">5. Elezione domicilio digitale e invio digitale</Typography>
+          <Typography variant="body1">
+            5.1 Se hai eletto un domicilio digitale di Piattaforma, a prescindere che tale domicilio sia stato da Te
+            eletto per un numero ristretto di Pubbliche Amministrazioni o per tutte le Pubbliche Amministrazione
+            utilizzatrici della Piattaforma, per le notificazioni digitali nei Tuoi confronti gli eventuali diversi
+            domicili (quello generale presente in un elenco pubblico, ad esempio quello iscritto nel Tuo albo
+            professionale di appartenenza ovvero quello eletto presso la singola Pubblica Amministrazione mittente)
+            verranno utilizzati laddove l’invio al Tuo domicilio digitale di Piattaforma risulti saturo, invalido,
+            inattivo, la Piattaforma è chiamata a tentare - decorsi 7 giorni dal primo invio - un secondo invio
+            digitale, utilizzando tutti i Tuoi domicili digitali disponibili, in aggiunta a quello di
+            Piattaforma.<br/>
+            5.2 Pertanto, fai attenzione: controlla che il Tuo indirizzo pec che hai inserito come Tuo domicilio
+            eletto o comunque che risulta presente in un elenco pubblico previsto dalla legge non sia sia saturo o
+            comunque verificane la sua operatività poiché potrebbe essere pregiudicata la ricezione in via digitale
+            delle notifiche nei Tuoi confronti con ogni possibile aggravio di costi per la notifica.<br/>
+            5.3 Infatti, se il domicilio digitale risulti essere, saturo, invalido e/o inattivo, riceverai il
+            documento oggetto di notifica in modalità analogica e la Piattaforma Ti darà evidenza di ciò con un
+            avviso di mancato recapito disponibile nella Tua area riservata, proprio per informarti del fatto che la
+            notifica procederà in via analogica.<br/>
+            5.4 Riceverai, infatti, a mezzo raccomandata semplice la notizia della notifica processata nei Tuoi
+            confronti.<br/>
+            5.5 Si procede secondo quanto descritto ai punti precedenti, ogni qualvolta la Piattaforma debba
+            processare una notifica nei confronti di un Utente che abbia una pluralità di domicili digitali e
+            risulti saturo, invalido e/o inattivo il domicilio digitale eletto dall’Utente.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">6. Recapito digitale e avviso di cortesia</Typography>
+          <Typography variant="body1">
+            6.1 Tramite la Piattaforma gli Utenti possono censire dei canali di comunicazione diversi dal domicilio
+            digitale, ad esempio, un numero di telefono cellulare o una e-mail ovvero usufruire del servizio di
+            messaggistica dell’app IO (di seguito <b>“Recapiti digitali”</b>), presso i quali ricevere gli avvisi di
+            cortesia.<br/>
+            6.2. Gli avvisi di cortesia vengono inviati su tutti i Recapiti digitali disponibili.<br/>
+            6.3 Ai fini del censimento dei Recapiti digitali riceverai sugli stessi un messaggio per validare
+            l’inserimento dei Recapiti Digitali diversi dall’app IO.<br/>
+            6.4 Se sei già un utente dell’app IO e hai già attivo il servizio di messaggistica la Piattaforma
+            utilizzerà tale canale per inviarti gli avvisi di cortesia, salvo la Tua facoltà di indicare altri
+            Recapito Digitali come indicato al punto 5.1.<br/>
+            6.5 Tali avvisi di cortesia, pur non avendo di per sé valore legale, Ti permettono di accedere tramite
+            link ai documenti informatici oggetto di notifica, anche prima di aver ricevuto la notifica e Ti
+            consentono di perfezionare la notifica attraverso l’accesso alla Tua area riservata con ogni possibile
+            risparmio in termini di tempo e di spese di notificazione.<br/>
+            6.6 Fai però attenzione, poiché se decorsi 7 giorni solari dalla data in cui il documento informatico
+            oggetto di notificazione è reso disponibile sulla Piattaforma, Tu non hai avuto accesso alla Tua area
+            riservata della Piattaforma, - per ottenere il perfezionamento della notifica nei Tuoi confronti - il
+            gestore della Piattaforma richiede all’addetto al recapito di provvedere alla notifica dell’avviso di
+            avvenuta ricezione in formato cartaceo, a mezzo posta, tramite raccomandata con ricevuta di ritorno
+            ovvero con le modalità previste dalla legge 20 novembre 1982, n. 890, con ogni possibile aggravio di
+            costi di notifica a Tuo carico (cfr. paragrafo 7).<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">7. Notificazione in modalità analogica</Typography>
+          <Typography variant="body1">
+            7.1 Se non hai nessun domicilio digitale (cfr. paragrafo 4) e/o non hai perfezionato la notifica
+            accedendo alla Tua area riservata della Piattaforma, anche a seguito della ricezione di un un avviso di
+            cortesia (cfr. paragrafo 5), la notifica nei Tuoi confronti deve essere processata in modalità analogica
+            attraverso l’invio cartaceo dell’avviso di avvenuta ricezione, con ogni possibile aggravio di costi di
+            notifica a Tuo carico.<br/>
+            7.2 Tale avviso, Ti verrà inviato tramite raccomandata con ricevuta di ritorno ovvero con le modalità
+            previste dalla legge 20 novembre 1982, n. 890,<br/>
+            7.3 Qualora l’Utente risultasse irreperibile nel corso dei tentativi di spedizione, l’Utente potrà
+            comunque acquisire il documento informatico oggetto di notificazione (cfr. par. 2.3).<br/>
+            7.4 Qualora Tu volessi maggiori dettagli sulla spedizione analogica rispetto a quelli resi disponibili
+            nella Tua area riservata della Piattaforma, devi rivolgerti all’addetto al recapito di riferimento della
+            spedizione.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">8. Costi e Spese di notificazione</Typography>
+          <Typography variant="body1">
+            8.1 L’utilizzo della Piattaforma non prevede alcun corrispettivo da parte degli Utenti nei confronti
+            della Società.<br/>
+            8.2 Tuttavia la Pubblica Amministrazione che esegue la notifica ha diritto di richiederTi il rimborso
+            delle spese di notifica da essa sostenute.<br/>
+            8.3 I costi del servizio sono quelli indicati nel Decreto Costi agli articoli 3 e 4.<br/>
+            8.4 Resta inteso che, oltre alle spese di notifica, potrà altresì esserTi richiesto l’importo
+            eventualmente previsto nell’atto notificato.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">9. Assistenza all’Utente e malfunzionamenti della
+            Piattaforma</Typography>
+          <Typography variant="body1">
+            9.1 In Piattaforma è reso altresì disponibile un servizio di assistenza per la risoluzione di ogni
+            eventuale anomalia ovvero per ogni richiesta di informazione, chiarimento e/o approfondimento
+            sull’operatività e sulle funzionalità della Piattaforma stessa o per segnalazioni relative
+            all’accessibilità.<br/>
+            9.2 Fai attenzione: ricorda di utilizzare il servizio di assistenza disponibile sulla Piattaforma anche
+            per segnalare problemi o anomalie nella visualizzazione dei documenti digitali oggetto di notificazione
+            e/o dei relativi allegati. Ti ricordiamo, infatti, che la notifica si perfeziona nei Tuoi confronti
+            anche laddove esistessero dei problemi o anomalie nella visualizzazione dei documenti che non siano
+            collegati ad un malfunzionamento della piattaforma.<br/>
+            9.3 I malfunzionamenti, ovvero i possibili disservizi della Piattaforma, sono segnalati sul Sito e con
+            le stesse modalità la Società comunica il ripristino della funzionalità della stessa.<br/>
+            9.4 Tramite la Piattaforma sono resi disponibili gli atti opponibili ai terzi, ovvero le attestazioni
+            relative al periodo di malfunzionamento della Piattaforma e quelle relative alla data di ripristino
+            delle funzionalità della stessa (cfr. par. 1.5).<br/>
+            9.5 Puoi segnalare un malfunzionamento al servizio di assistenza il quale provvede, nel caso lo stesso
+            non sia già stato individuato attraverso i sistemi di monitoraggio automatico, a gestire lo stesso.<br/>
+            9.6 La Società, durante il periodo di malfunzionamento, non è tenuta a fornire alcun tipo di assistenza
+            diverso o ulteriore rispetto a quanto indicato nel presente articolo.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">10. Responsabilità e obblighi dell’Utente</Typography>
+          <Typography variant="body1" component="span">
+            10.1 Utilizzando la Piattaforma, accetti i termini e le condizioni dei presenti ToS e ti impegni a
+            rispettarli in ogni momento. Puoi utilizzare la Piattaforma esclusivamente per gli scopi autorizzati dai
+            presenti ToS e in conformità alla legge applicabile.<br/>
+            10. 2 Ti impegni, in particolare, a non utilizzare, direttamente o indirettamente, la Piattaforma:<br/>
+            <ul style={{listStyle: "lower-roman"}}>
+              <li>sfruttando l’identità di un’altra persona o ricorrendo all’utilizzo non personale delle
+                funzionalità della Piattaforma, salvo quando eventualmente ed espressamente previsto dai presenti
+                ToS;
+              </li>
+              <li>per comunicare dati o informazioni false, ingannevoli o illecite;</li>
+              <li>per trasmettere virus, malware o altri codici dannosi per qualsivoglia dispositivo o sistema;</li>
+              <li>in violazione dei diritti della Società o di terzi;</li>
+              <li>per finalità o scopi commerciali;</li>
+              <li>in modo illecito, diffamatorio, osceno, volgare, intimidatorio, offensivo nei confronti della
+                Società o di terzi o contrario a norme imperative, all’ordine pubblico e al buon costume;
+              </li>
+            </ul>
+            per trasmettere comunicazioni illecite o non autorizzate, quali messaggi massivi, spam o messaggi
+            automatici.
+            10.3 Ogni attività effettuata sulla Piattaforma dagli Utenti è eseguita in nome proprio e con ogni
+            conseguenza ai fini di legge, anche in merito ad eventuali responsabilità per eventuali perdite o danni
+            subiti dalla Società o da terzi in caso di violazioni dei presenti ToS, nonché dell’utilizzo improprio o
+            illecito della Piattaforma.<br/>
+            10.4 La Società può adottare autonomamente tutte le misure necessarie, inclusa la sospensione o
+            l’interruzione dei servizi erogati dalla Piattaforma e/o inibendo l’utilizzo della stessa da parte
+            dell’Utente, qualora ne faccia un utilizzo illecito o altrimenti vietato, oppure commetta una violazione
+            sostanziale o reiterata dei presenti ToS.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">11. Esclusioni e limitazione di responsabilità della
+            Società</Typography>
+          <Typography variant="body1">
+            11.1 Sei il solo responsabile dei dati inseriti per accedere alla Piattaforma.<br/>
+            11.2 In considerazione di quanto precisato, la Società è responsabile unicamente del corretto
+            funzionamento dei Servizi erogati tramite la Piattaforma, fatte salve le responsabilità in capo a
+            soggetti terzi, quali ad esempio l’addetto al recapito o FSU, per le attività di rispettiva
+            competenza.<br/>
+            11.3 La Società non potrà quindi essere considerata responsabile per nessun danno, perdita, pregiudizio,
+            che dovesse derivare dall’utilizzo della Piattaforma per attività non di sua competenza ai sensi del
+            Decreto e del Decreto Funzionamento.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">12. Privacy</Typography>
+          <Typography variant="body1">
+            12.1 Per utilizzare la Piattaforma è necessario, prima di accedere alla stessa, prendere visione
+            dell’Informativa sul trattamento dei dati personali.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">13. Modifiche ai ToS </Typography>
+          <Typography variant="body1">
+            13.1 La Società si riserva il diritto di modificare, anche in forza di nuove previsioni di legge, in
+            qualsiasi momento e anche senza preavviso, in tutto o in parte, i presenti ToS, dandone adeguata
+            pubblicità all’Utente.<br/>
+            13.2 Nuovi Termini e Condizioni saranno efficaci a far data dalla pubblicazione sul Sito, e non saranno
+            mai applicati retroattivamente.<br/>
+            13.3 Resta inteso che l’utilizzo della Piattaforma o dei Servizi in seguito all’entrata in vigore delle
+            modifiche ai ToS comporta l’accettazione dei medesimi da parte dell’Utente.<br/>
+            13.4 L’eventuale nullità, annullabilità o inefficacia di una o più clausole dei ToS non si estende alle
+            altre clausole, che ove possibile conservano validità.<br/>
+            13.5 Se non intendi accettare le modifiche apportate ai presenti ToS, come unico rimedio, potrai, in
+            qualsiasi momento, senza alcun preavviso, cessare l’utilizzo della Piattaforma stessa.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1" fontWeight="bold">14. Legge applicabile e foro competente</Typography>
+          <Typography variant="body1">
+            14.1 I presenti ToS sono regolati dalla legge italiana.<br/>
+            14.2 Ogni controversia che dovesse insorgere tra l’Utente e la Società in relazione ai presenti ToS e/o
+            alla Piattaforma sarà di competenza esclusiva del foro di Roma oppure del giudice del luogo di residenza
+            o domicilio dell’Utente, ove qualificato come consumatore ai sensi del D.lgs. 206/2005.<br/>
+          </Typography>
+          <br/>
+          <Typography variant="body1">
+            Ai sensi degli articoli 1341 e 1342 del codice civile, proseguendo l’Utente dichiara espressamente di
+            aver letto, compreso e accettato i seguenti articoli dei presenti termini e condizioni
+            d’uso: <b>(8)</b> Costi e Spese; <b>(9)</b> Assistenza all’Utente e malfunzionamenti della
+            Piattaforma; <b>10)z</b> Responsabilità e obblighi dell’Utente; <b>(11)</b> Esclusioni e limitazione di
+            responsabilità della Società; <b>(13)</b> Modifiche ai ToS e <b>(14)</b> Legge applicabile e foro
+            competente.
+          </Typography>
+        </Box>
+      </Container>
+    </>);
 
-  const handleAccept = () => {
-    void dispatch(acceptToS()).unwrap()
-    .then(() => {
-      navigate(routes.NOTIFICHE);
-    })
-    .catch(_ => {
-      console.error(_);
-    });
-  };
-
-  useEffect(() => {
-    if (tos) {
-      navigate(routes.NOTIFICHE);
-    }
-  }, [tos]);
-
-  return (
-    <Fragment>
-      <Grid container justifyContent="center" my={isMobile ? 4 : 16}>
-        <Grid item xs={10} sm={8} md={4} display="flex" alignItems="center" flexDirection="column">
-          <TermsOfServiceHandler
-            handleAcceptTos={handleAccept}
-            transComponent={(path: string, components: Array<ReactNode>, defaultLoaclization: ReactNode) => (
-              <Trans i18nKey={path} components={components}>
-                {defaultLoaclization}
-              </Trans>
-            )}
-          />
-        </Grid>
-      </Grid>
-    </Fragment>
-  );
-};
-
-export default TermsOfService;
+export default PrivacyPolicyPage;
